mock_band working with tester.py
