+++ contracts/mock_band/README.md
@@ -1,0 +1,21 @@
+# Mock Band Contract
+* [Introduction](#Introduction)
+* [Sections](#Sections)
+    * [User](#User)
+        * Queries
+            * [GetReferenceData](#GetReferenceData)
+# Introduction
+The Mocked Band contract is used to test locally when there is no official band contract available
+
+### Queries
+
+#### GetReferenceData
+Get a hardcoded sample from an ETH query for testing locally
+##### Response
+```json
+{
+  "rate": "3119154999999000000000",
+  "last_updated_base": 1628548483,
+  "last_updated_quote": 3377610
+}
+```
